refactor: внести логику GuardService-ов обратно в основные сервисы

+++ main-svc/src/main/java/ru/practicum/explorewithme/mainsvc/category/service/CategoryServiceImpl.java
@@ -9,9 +9,11 @@ import ru.practicum.explorewithme.mainsvc.category.dto.CategoryDto;
 import ru.practicum.explorewithme.mainsvc.category.entity.Category;
 import ru.practicum.explorewithme.mainsvc.category.mapper.CategoryMapper;
 import ru.practicum.explorewithme.mainsvc.category.repository.CategoryRepository;
-import ru.practicum.explorewithme.mainsvc.category.util.CategoryGuardService;
 import ru.practicum.explorewithme.mainsvc.common.requests.PaginationRequest;
 import ru.practicum.explorewithme.mainsvc.common.utils.pageable.PageableUtility;
+import ru.practicum.explorewithme.mainsvc.exception.AccessRightsException;
+import ru.practicum.explorewithme.mainsvc.exception.AlreadyExistsException;
+import ru.practicum.explorewithme.mainsvc.exception.EntityNotFoundException;
 
 import java.util.List;
 
@@ -20,14 +22,16 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryMapper categoryMapper;
-    private final CategoryGuardService categoryExceptionChecker;
     private final CategoryRepository categoryRepository;
     private final PageableUtility pageableUtility;
 
     @Transactional
     @Override
     public CategoryDto addCategory(CategoryDto categoryDto) {
-        categoryExceptionChecker.checkNameUniqueness(categoryDto.getName());
+        String name = categoryDto.getName();
+        if (categoryRepository.existsByName(name)) {
+            throw new AlreadyExistsException("Category with name = " + name + " already exists.");
+        }
 
         Category category = categoryMapper.toEntity(categoryDto);
         Category savedCategory = categoryRepository.save(category);
@@ -40,8 +44,12 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     @Override
     public void deleteCategory(Long catId) {
-        categoryExceptionChecker.checkExistenceById(catId);
-        categoryExceptionChecker.checkNotContainEvents(catId);
+        if (!categoryRepository.existsById(catId)) {
+            throw new EntityNotFoundException("Category with id = " + catId + " not found.");
+        }
+        if (categoryRepository.existsEventByCategoryId(catId)) {
+            throw new AccessRightsException("Category with id = " + catId + " has events.");
+        }
         categoryRepository.deleteById(catId);
         log.info("Category with id = {} has been deleted.", catId);
     }
@@ -49,11 +57,13 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     @Override
     public CategoryDto patchCategory(Long catId, CategoryDto categoryDto) {
-        Category category = categoryExceptionChecker.findById(catId);
+        Category category = this.findCategoryById(catId);
 
         String name = categoryDto.getName();
         if (!name.equals(category.getName())) {
-            categoryExceptionChecker.checkNameUniqueness(name);
+            if (categoryRepository.existsByName(name)) {
+                throw new AlreadyExistsException("Category with name = " + name + " already exists.");
+            }
             category.setName(name);
         }
 
@@ -67,7 +77,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional(readOnly = true)
     @Override
     public CategoryDto getCategoryById(Long catId) {
-        Category category = categoryExceptionChecker.findById(catId);
+        Category category = this.findCategoryById(catId);
         CategoryDto dto = categoryMapper.toDto(category);
         log.info("Category with id = {} has been found. Category : {}", catId, dto);
         return dto;
@@ -82,5 +92,13 @@ public class CategoryServiceImpl implements CategoryService {
         List<CategoryDto> dtos = categoryMapper.toDtoList(categories);
         log.info("Categories have been found. List size : {}", dtos.size());
         return dtos;
+    }
+
+    @Transactional(readOnly = true)
+    @Override
+    public Category findCategoryById(long id) {
+        return categoryRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException("Category with id = " + id + " not found.")
+        );
     }
 }
